Done refactoring code. Debugging and testing will happen tommorrow, and implementing merging should be a peach of cake at that point.
